requirements.txt
Django~=2.2
django-bootstrap4
django-simple-captcha
